move esp_loader.ts to correct folder
